Create creation of linkedlist
This Node class represents individual nodes in a linked list. Each node has two attributes:
value: It stores the data or value associated with the node.
next: It's initially set to None and will later reference the next node in the linked list.
his LinkedList class represents the linked list itself. It has the following attributes:
head: This is initially set to the first node created with the provided value.
tail: This is also initially set to the first node.
length: It keeps track of the number of nodes in the linked list, which is initially set to 1 since there is one node created during initialization.